[SPARK-57288][SQL][TESTS] Reorganize MERGE INTO schema evolution tests by scenario
Reorganize the MERGE INTO schema evolution test traits so each test lives
in a file named after the schema evolution scenario it exercises, instead
of the previous top-level vs. nested split.

The new traits are:
- MergeIntoSchemaEvolutionExtraSourceColumnTests
- MergeIntoSchemaEvolutionExtraTargetColumnTests
- MergeIntoSchemaEvolutionTypeWideningTests
- MergeIntoSchemaEvolutionMixedTests

The total set of tests is unchanged.
